requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI
openai>=1.0.0

# PDF processing
pdfplumber>=0.11.0
python-docx>=1.1.0
reportlab>=4.0.0

# Web framework
streamlit>=1.55.0

# Optional: API backend (FastAPI)
# fastapi>=0.104.0
# uvicorn[standard]>=0.24.0

# Development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
